Update FAQ entry for default_flow_style to reflect PyYAML 5.1 changes
This behavior led to some confusion recently as we have some boxes
running PyYAML 5.3 and some running older PyYAML. Code originally
written for 5.3 produced different output on older releases.

This change adds clarification to the FAQ entry for default_flow_style
which mentions that the default behavior changed in version 5.1.